# docs/api/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Sensei++ API",
    "description": "Versioned API. Owner-scoped operations require X-Owner-Id during development; this header selects an owner and is not authentication. Production authentication is not implemented yet. Resources return an opaque versionToken and an ETag header. Send that ETag as a single strong If-Match header on updates, approvals, and DELETE actions. Missing If-Match returns 428, malformed tokens return 400, and stale tokens return 412. Creates do not require If-Match and return the initial ETag. Lists return { items, nextCursor }; limit defaults to 25 and accepts 1 to 100. Pass nextCursor as cursor with the same owner and filters. Errors use application/problem+json with code and traceId.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/evidence/knowledge": {
      "get": {
        "tags": [
          "Evidence and Profile"
        ],
        "operationId": "ListKnowledgeSummaries",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size from 1 to 100; defaults to 25.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque nextCursor from the preceding page. Keep owner and filters unchanged.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Owner-Id",
            "in": "header",
            "description": "Required development owner selector (UUID). This is not authentication.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageEnvelopeOfKnowledgeResource"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/knowledge/{id}": {
      "get": {
        "tags": [
          "Evidence and Profile"
        ],
        "operationId": "GetKnowledgeSummary",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Owner-Id",
            "in": "header",
            "description": "Required development owner selector (UUID). This is not authentication.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeResource"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/learning/topics": {
      "get": {
        "tags": [
          "Learning runtime"
        ],
        "operationId": "BrowseLearningTopics",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parent",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size from 1 to 100; defaults to 25.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque nextCursor from the preceding page. Keep owner and filters unchanged.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageEnvelopeOfTopicResource"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learning/topics/{id}": {
      "get": {
        "tags": [
          "Learning runtime"
        ],
        "operationId": "GetLearningTopic",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicResource"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learning/concept-relations": {
      "post": {
        "tags": [
          "Learning runtime"
        ],
        "operationId": "CreateConceptRelation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConceptRelation"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learning/materials/{id}": {
      "get": {
        "tags": [
          "Learning runtime"
        ],
        "operationId": "GetLearningMaterial",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResource"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learning/material-views": {
      "post": {
        "tags": [
          "Learning runtime"
        ],
        "operationId": "RecordMaterialView",
        "parameters": [
          {
            "name": "X-Owner-Id",
            "in": "header",
            "description": "Required development owner selector (UUID). This is not authentication.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityReceipt"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learning/session-previews": {
      "post": {
        "tags": [
          "Learning runtime"
        ],
        "operationId": "PreviewLearningSession",
        "parameters": [
          {
            "name": "X-Owner-Id",
            "in": "header",
            "description": "Required development owner selector (UUID). This is not authentication.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionSetup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preview"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details with code and traceId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KnowledgeResource": {
        "required": [
          "conceptId",
          "familyCount",
          "assistedCount",
          "selfReviewCount",
          "qualifyingFamilies",
          "qualifyingSessions",
          "correctness",
          "observationIds",
          "nextPositiveReviewAt"
        ],
        "type": "object",
        "properties": {
          "conceptId": {
            "type": "string",
            "format": "uuid"
          },
          "familyCount": {
            "type": "integer",
            "format": "int32"
          },
          "assistedCount": {
            "type": "integer",
            "format": "int32"
          },
          "selfReviewCount": {
            "type": "integer",
            "format": "int32"
          },
          "qualifyingFamilies": {
            "type": "integer",
            "format": "int32"
          },
          "qualifyingSessions": {
            "type": "integer",
            "format": "int32"
          },
          "correctness": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "observationIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "nextPositiveReviewAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "MaterialViewRequest": {
        "required": [
          "operationId",
          "materialVersionId"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "materialVersionId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "PageEnvelopeOfKnowledgeResource": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KnowledgeResource"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          },
          "traceId": {
            "type": "string",
            "description": "Request trace identifier for diagnostics."
          }
        }
      },
...
    }
  },
...
}